+++ typescript/packages/data-contracts-expo-sqlite/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @baukit/data-contracts-expo-sqlite
 
+## 0.3.0
+
+### Minor Changes
+
+- Release the coordinated baukit 0.3.0 train.
+
+### Patch Changes
+
+- Updated dependencies
+  - @baukit/data-contracts@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes