+++ typescript/packages/data-contracts/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @baukit/data-contracts
 
+## 0.3.0
+
+### Minor Changes
+
+- Release the coordinated baukit 0.3.0 train.
+
 ## 0.2.0
 
 ### Minor Changes